added input

+++ Assets/Scripts/PlayerView.cs
@@ -2,7 +2,7 @@ using UnityEngine;
 
 public class PlayerView : MonoBehaviour
 {
-    public void SetPosition(int pos)
+    public void SetPosition(float pos)
     {
         transform.position = new Vector3(0,pos,0);
     }
